Convert to standard gradle project structure
