+++ .gitignore
@@ -1,0 +1,6 @@
+# Gradle files
+.gradle/*
+build/*
+
+# IDE files
+.idea/*
