+++ build.gradle
@@ -1,48 +1,48 @@
 plugins {
     id 'java-library'
-    id 'eclipse'
 }
 
-version = '1.0'
+version = '1.0.0'
 sourceCompatibility = 1.8
 targetCompatibility = 1.8
-
-sourceSets {
-    main.java.srcDirs = ['src']
-    main.resources.srcDirs = ['src']
-}
 
 compileJava.options.encoding = 'UTF-8'
 compileTestJava.options.encoding = 'UTF-8'
 javadoc.options.encoding = 'UTF-8'
- 
+
+ext {
+    asm_ver = "8.0.1"
+    mainClassName = "me.nov.dalvikgate.DexToASM"
+}
+
 repositories {
     mavenCentral()
 }
 
 dependencies {
-	compile 'commons-io:commons-io:2.6' 
+    compile "org.ow2.asm:asm:${asm_ver}"
+    compile "org.ow2.asm:asm-tree:${asm_ver}"
+    compile "org.ow2.asm:asm-analysis:${asm_ver}"
+    compile "org.ow2.asm:asm-util:${asm_ver}"
 
-	compile 'org.ow2.asm:asm:8.0.1'
-	compile 'org.ow2.asm:asm-tree:8.0.1'
-	compile 'org.ow2.asm:asm-analysis:8.0.1'
-	compile 'org.ow2.asm:asm-util:8.0.1'
-	
-	compile 'org.smali:dexlib2:2.4.0'
+    compile 'commons-io:commons-io:2.6'
+
+    compile 'org.smali:dexlib2:2.4.0'
 }
 
 task fatJar(type: Jar) {
-	manifest {
-        attributes 'Implementation-Title': 'dalvikgate',  
-        	'Implementation-Version': version,
-        	'Main-Class': 'me.nov.dalvikgate.DexToASM'
+    manifest {
+        attributes 'Implementation-Title': "${project.name}",
+                'Implementation-Version': "${project.version}",
+                'Main-Class': "${mainClassName}"
     }
-    baseName = project.name + '-all'
-    from (configurations.compile.collect { entry -> zipTree(entry) }) {
+    baseName(project.name + '-all')
+    from(configurations.compile.collect { entry -> zipTree(entry) }) {
         exclude 'META-INF/MANIFEST.MF'
         exclude 'META-INF/*.SF'
         exclude 'META-INF/*.DSA'
         exclude 'META-INF/*.RSA'
+        exclude 'module-info.*'
     }
     with jar
 }
